# Conflicts:
#	UltraEmeraldScriptEditor/CompileSupport/CompileSupport.csproj

+++ UltraEmeraldScriptEditor/EditorSupport/Undo/UndoStack.cs
@@ -11,13 +11,26 @@ namespace EditorSupport.Undo
         {
             _operations = new List<IUndoableOperation>();
             _undoProcess = 0;
+            _undoing = false;
+            _group = null;
         }
 
         public virtual void AddOperation(IUndoableOperation operation)
         {
-            _operations.RemoveRange(_undoProcess, _operations.Count - _undoProcess);
-            _operations.Add(operation);
-            ++_undoProcess;
+            if (_undoing)
+            {
+                return;
+            }
+            if (_group != null)
+            {
+                _group.Operations.Add(operation);
+            }
+            else
+            {
+                _operations.RemoveRange(_undoProcess, _operations.Count - _undoProcess);
+                _operations.Add(operation);
+                ++_undoProcess;
+            }
         }
 
         public virtual Boolean Undo()
@@ -26,8 +39,10 @@ namespace EditorSupport.Undo
             {
                 return false;
             }
+            _undoing = true;
             _operations[_undoProcess - 1].Undo();
             --_undoProcess;
+            _undoing = false;
             return true;
         }
 
@@ -37,8 +52,10 @@ namespace EditorSupport.Undo
             {
                 return false;
             }
+            _undoing = true;
             _operations[_undoProcess].Redo();
             ++_undoProcess;
+            _undoing = false;
             return true;
         }
 
@@ -58,7 +75,29 @@ namespace EditorSupport.Undo
             return _undoProcess < _operations.Count;
         }
 
+        public void StartGrouping()
+        {
+            if (_group != null)
+            {
+                throw new InvalidOperationException("UndoStack has already started.");
+            }
+            _group = new UndoOperationGroup();
+        }
+
+        public void EndGrouping()
+        {
+            if (_group == null)
+            {
+                throw new InvalidOperationException("UndoStack hasn't started yet.");
+            }
+            var group = _group;
+            _group = null;
+            AddOperation(group);
+        }
+
         protected List<IUndoableOperation> _operations;
+        protected UndoOperationGroup _group;
+        protected Boolean _undoing;
         protected Int32 _undoProcess;
     }
 }
